# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SprintPulse API
  version: 0.2.0
  license:
    name: Private hackathon prototype
  description: |
    OpenAPI contract for the Express routes mounted at `/api` in SprintPulse.

    The API currently uses a `personaId` query or body field to resolve the acting user. Some routes are available only
    when `ENABLE_MOCK_FLOW=true`; in Supabase data mode those routes return `501` until their adapters are implemented.
servers:
  - url: http://localhost:4000/api
    description: Local SprintPulse API
...
security: []
paths:
...
  /projects/{projectId}/team:
    get:
      tags: [Team]
      operationId: getProjectTeam
      summary: Get project team members and pending invites.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
        - $ref: "#/components/parameters/PersonaIdQuery"
      responses:
        "200":
          description: Team membership and edit capability.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
        "404":
          description: Team is not found or not visible.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notVisible:
                  value:
                    error: Team not found or not visible to this user
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Team]
      operationId: addProjectMemberViaTeam
      summary: Add or invite a project member.
      description: Alias of `POST /projects/{projectId}/invites`. In mock flow this route returns `501`.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        $ref: "#/components/requestBodies/InviteProjectMemberBody"
      responses:
        "201":
          description: Project member invited or linked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteProjectMemberResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /projects/{projectId}/invites:
    post:
      tags: [Team]
      operationId: inviteProjectMember
      summary: Invite or upsert a project member.
      description: In mock flow this route returns `501`.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
      requestBody:
        $ref: "#/components/requestBodies/InviteProjectMemberBody"
      responses:
        "201":
          description: Project member invited or linked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteProjectMemberResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /projects/{projectId}/team/{profileId}:
    patch:
      tags: [Team]
      operationId: updateProjectMember
      summary: Update a project member's role or integration handles.
      description: In mock flow this route returns `501`.
      parameters:
        - $ref: "#/components/parameters/ProjectIdPath"
        - $ref: "#/components/parameters/ProfileIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectMemberRequest"
            examples:
              updateMember:
                value:
                  personaId: priya
                  role: scrum-master
                  jiraAccountId: himanshu.atlassian
                  githubUsername: himanshu-sp
      responses:
        "200":
          description: Updated team response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
        "404":
          description: Team member is not found or not visible.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notVisible:
                  value:
                    error: Team member not found or not visible to this user
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
...
components:
  parameters:
    PersonaIdQuery:
      name: personaId
      in: query
      required: true
      description: Persona/profile id used by SprintPulse to resolve viewer permissions.
      schema:
        type: string
      example: priya
...
    ProjectIdPath:
      name: projectId
      in: path
      required: true
      description: Project id or, in mock flow, case-insensitive project key.
      schema:
        type: string
      example: sp-core
...
    ProfileIdPath:
      name: profileId
      in: path
      required: true
      description: Profile id of the project member to update.
      schema:
        type: string
      example: himanshu
  requestBodies:
    InviteProjectMemberBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InviteProjectMemberRequest"
          examples:
            inviteDeveloper:
              value:
                personaId: priya
                email: dev@sprintpulse.dev
                name: Dev Sharma
                appRole: developer
                projectRole: developer
                title: Full-stack Engineer
                jiraAccountId: dev.atlassian
                githubUsername: dev-sp
...
  responses:
    InternalServerError:
      description: The route failed while loading or mutating data.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internal:
              value:
                error: Unable to load data
...
    NotImplemented:
      description: Route is not available in the current runtime mode.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notImplemented:
              value:
                error: Supabase data mode is enabled, but this route still needs its database-backed adapter.
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Persona not found
...
    HackathonRole:
      type: string
      enum: [frontend, backend, architect, qa]
    ProductPersona:
      type: string
      enum: [product-owner, engineering-manager, scrum-master, developer, qa-lead, presenter]
    AccessScope:
      type: string
      enum: [team, individual, quality, presentation]
    AppRole:
      type: string
      enum: [admin, product-owner, engineering-manager, scrum-master, developer, qa-lead]
    Permission:
      type: string
      enum:
        - project:view
        - project:create
        - project:connect
        - project:editTeam
        - standup:submit
        - standup:sync
        - dashboard:viewTeam
        - dashboard:viewOwn
        - member:viewTeam
        - member:viewOwn
    ProjectRole:
      type: string
      enum: [product-owner, scrum-master, engineering-manager, architect, developer, qa]
    ProjectSource:
      type: string
      enum: [manual, jira]
...
    Persona:
      type: object
      required: [id, name, email, initials, hackathonRole, productPersona, title, accessScope, focus]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        initials:
          type: string
        hackathonRole:
          $ref: "#/components/schemas/HackathonRole"
        productPersona:
          $ref: "#/components/schemas/ProductPersona"
        title:
          type: string
        accessScope:
          $ref: "#/components/schemas/AccessScope"
        focus:
          type: string
...
    UserProfile:
      type: object
      required: [id, email, name, initials, title, appRole, productPersona, accessScope, status, createdAt]
      properties:
        id:
          type: string
        authUserId:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        initials:
          type: string
        title:
          type: string
        appRole:
          $ref: "#/components/schemas/AppRole"
        productPersona:
          $ref: "#/components/schemas/ProductPersona"
        accessScope:
          $ref: "#/components/schemas/AccessScope"
        status:
          type: string
          enum: [active, invited]
        createdAt:
          type: string
          format: date-time
        invitedBy:
          type: string
...
    ProjectMember:
      type: object
      required: [personaId, name, email, initials, role]
      properties:
        personaId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        initials:
          type: string
        role:
          $ref: "#/components/schemas/ProjectRole"
        jiraAccountId:
          type: string
        githubUsername:
          type: string
    SprintInfo:
      type: object
      required: [id, name, goal, startDate, endDate, status]
      properties:
        id:
          type: string
        name:
          type: string
        goal:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [planned, active, closed]
    SprintProject:
      type: object
      required: [id, key, name, source, sprint, members, ownerIds, scrumMasterIds, createdBy, createdAt, updatedAt]
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        source:
          $ref: "#/components/schemas/ProjectSource"
        jiraSite:
          type: string
        sprint:
          $ref: "#/components/schemas/SprintInfo"
        members:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMember"
        ownerIds:
          type: array
          items:
            type: string
        scrumMasterIds:
          type: array
          items:
            type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastSyncAt:
          type: string
          format: date-time
...
    ProjectInvite:
      type: object
      required: [id, projectId, email, role, invitedBy, status, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/ProjectRole"
        invitedBy:
          type: string
        status:
          type: string
          enum: [pending, accepted, expired]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    TeamResponse:
      type: object
      required: [viewer, project, permissions, members, invites, canEditTeam]
      properties:
        viewer:
          $ref: "#/components/schemas/Persona"
        project:
          $ref: "#/components/schemas/SprintProject"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        members:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMember"
        invites:
          type: array
          items:
            $ref: "#/components/schemas/ProjectInvite"
        canEditTeam:
          type: boolean
    InviteProjectMemberRequest:
      type: object
      required: [personaId, email, name, appRole, projectRole]
      properties:
        personaId:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        appRole:
          $ref: "#/components/schemas/AppRole"
        projectRole:
          $ref: "#/components/schemas/ProjectRole"
        title:
          type: string
        jiraAccountId:
          type: string
        githubUsername:
          type: string
    InviteProjectMemberResponse:
      type: object
      required: [profile, member, invite, warnings]
      properties:
        profile:
          $ref: "#/components/schemas/UserProfile"
        member:
          $ref: "#/components/schemas/ProjectMember"
        invite:
          $ref: "#/components/schemas/ProjectInvite"
        warnings:
          type: array
          items:
            type: string
        linkedSyncs:
          type: array
          description: Parallel evidence sync results for the other project signal sources.
          items:
            $ref: "#/components/schemas/ProjectSignalSyncStatus"
    UpdateProjectMemberRequest:
      type: object
      required: [personaId]
      properties:
        personaId:
          type: string
        role:
          $ref: "#/components/schemas/ProjectRole"
        jiraAccountId:
          type: string
        githubUsername:
          type: string
...
    ProjectSignalSyncStatus:
      type: object
      required: [source, status]
      properties:
        source:
          type: string
          enum: [standup, jira, git]
        status:
          type: string
          enum: [succeeded, failed]
        importedCount:
          type: integer
          minimum: 0
        warning:
          type: string
...